Backend/servicios-vision-artificial/matriz-imagen-api-ms/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
Pillow
numpy
opencv-python
matplotlib
httpx
